Reorganize package edit and show components
