+++ src/components/package/edit-managed/managed-package-edit.css
