+++ src/components/package/show/package-show.css
